Organized, fixed errors with imports, Added more to-do, attempted to remove all emojis
