--- .github/FUNDING.yml
@@ -1,4 +1,0 @@
-# These are supported funding model platforms
-
-github: theshadow76
-custom: ['https://paypal.me/shadowtechsc?country.x=CL&locale.x=es_XC]'
