--- TODO/todo.md
@@ -1,6 +1,0 @@
-# TO-DO List
-
-### Add Template for a basic PO bot
- - TBD
-### Organize files more efficiently
- - TBD
