+++ requirements.txt
@@ -2,9 +2,12 @@ aiohttp>=3.8.0
 websockets>=11.0.0
 asyncio
 python-dotenv>=1.0.0
-pandas>=1.5.0
 tzlocal>=4.0.0
-pydantic>=2.0.0
 typing-extensions>=4.0.0
-loguru>=0.7.0
 rich>=13.0.0
+selenium>=4.0.0
+webdriver-manager>=4.0.0
+psutil>=5.9.0
+loguru>=0.7.2
+pydantic>=2.0.0
+pandas>=2.0.0
