+++ todo.md
@@ -1,0 +1,7 @@
+# TO-DO List
+
+### Add Template for a basic PO bot
+ - Soon- Preferably before July 4th.
+
+### Integrate more fallbacks if there is any errors
+ - No ETA
