+++ tools/get_ssid.py
